Fixed import cycle due to parser usage in lang tests
