+++ lang/resolvers_test.go
@@ -1,0 +1,6 @@
+package lang_test
+
+type ImportExpectations struct {
+	filePath string
+	imports  []string
+}
