+++ pkg/test/helpers.go
@@ -10,7 +10,7 @@ import (
 )
 
 // SetupBasicPluginContext sets up a basic plugin context for testing plugins.
-func SetupBasicPluginContext(filePaths []string, languageCode core.LanguageCode) (core.TreeWalker, core.ImportAwareFileSystem, error) {
+func SetupBasicPluginContext(filePaths []string, languageCodes []core.LanguageCode) (core.TreeWalker, core.ImportAwareFileSystem, error) {
 	fileSystem, err := fs.NewLocalFileSystem(fs.LocalFileSystemConfig{
 		AppDirectories: filePaths,
 	})
@@ -19,17 +19,20 @@ func SetupBasicPluginContext(filePaths []string, languageCode core.LanguageCode)
 		return nil, nil, fmt.Errorf("failed to create file system: %w", err)
 	}
 
-	language, err := lang.GetLanguage(string(languageCode))
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get language: %w", err)
+	var languages []core.Language
+	for _, code := range languageCodes {
+		language, err := lang.GetLanguage(string(code))
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to get language: %w", err)
+		}
+		languages = append(languages, language)
 	}
-
-	walker, err := fs.NewSourceWalker(fs.SourceWalkerConfig{}, language)
+	walker, err := fs.NewSourceWalker(fs.SourceWalkerConfig{}, languages)
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to create source walker: %w", err)
 	}
 
-	treeWalker, err := parser.NewWalkingParser(walker, language)
+	treeWalker, err := parser.NewWalkingParser(walker)
 	if err != nil {
 		return treeWalker, nil, fmt.Errorf("failed to create tree walker: %w", err)
 	}
